+++ client/app/page.tsx
@@ -1,6 +1,6 @@
 import Header from "@/components/Header";
 import Sidebar from "@/components/Sidebar";
-import TableView from "@/components/TableView";
+import TableContainer from "./features/TableView/components/TableContainer";
 
 export default function Home() {
   return (
@@ -9,7 +9,7 @@ export default function Home() {
         <Sidebar />
         <div className="flex flex-1 flex-col overflow-hidden">
           <Header />
-          <TableView />
+          <TableContainer />
         </div>
       </div>
     </div>
